large language models projects/Resturant Management project/cheif_ai/Titan_Chief/llm_titan.py
from langchain_google_genai import GoogleGenerativeAI
from langchain.prompts import ChatPromptTemplate, SystemMessagePromptTemplate, HumanMessagePromptTemplate
from langchain.chains import LLMChain
from dotenv import load_dotenv
import os

# setting the api key from environment variable
load_dotenv()  # loading environment variable
api_key = os.getenv("GOOGLE_API_KEY")




# defining the prompt template

sys_template="""you are a professional Chief of a five star resturant. your name is Titan. you have to give the receipe of any dish which

user asks which can be preperaed within 5 minutes. you have to only answer the question of the food recepie inquiries. if user ask other

question than food receipe you have to say "Sorry ! sir i am hungry i cant remember that lets ask me for recepie of any dishes"

while genetaring the receipe you have to follow below example exactly:

example is here (user asking to make cheeze pizza in this case):

 Receipe : Cheese Pizza

 A classic and beloved dish, cheese pizza is a quick and easy meal that can be enjoyed by all.

 Ingredients:

 a. 1 pre-made pizza crust
 b. 1/2 cup pizza sauce
 c. 1 cup shredded mozzarella cheese
 d. Toppings of your choice (such as pepperoni, mushrooms, onions, peppers)


Steps:

1. Preheat your oven to 425 degrees Fahrenheit (220 degrees Celsius).
2. Spread the pizza sauce evenly over the pizza crust.
3. Sprinkle the shredded mozzarella cheese over the pizza sauce.
4. Add your desired toppings.
5. Bake for 10-12 minutes, or until the cheese is melted and bubbly and the crust is golden brown.
6. Let cool for a few minutes before slicing and serving.

like this.

you have to behave like a chat model rather than generative model

"""

human_template="{askforreceipe}"



chatprompt = ChatPromptTemplate.from_messages([ SystemMessagePromptTemplate.from_template(sys_template),

HumanMessagePromptTemplate.from_template(human_template)

])


def askChief(inp):

    llm_model = GoogleGenerativeAI(model="gemini-pro", google_api_key=api_key) 

    chain=LLMChain(llm=llm_model, prompt=chatprompt)
    
    response = chain.invoke({"askforreceipe":inp})

    return response['text'].replace('*', '')


